Split up Runner and List, so no need to connect to database in init
I want to define the migrations in init() functions, but I don't
necessarily want to connect to the databse in an init() function.
Decoupling the Runner and List from each other fixes this.

List is a simple list of all migrations that can be run.
Runner handles all logic related to actually running the migrations.